+++ cmd/internal/shared/applicationserver/config.go
@@ -20,8 +20,8 @@ import (
 
 	"go.thethings.network/lorawan-stack/cmd/internal/shared"
 	"go.thethings.network/lorawan-stack/pkg/applicationserver"
+	"go.thethings.network/lorawan-stack/pkg/applicationserver/io/web"
 	"go.thethings.network/lorawan-stack/pkg/config"
-	"go.thethings.network/lorawan-stack/pkg/webui"
 )
 
 // DefaultApplicationServerConfig is the default configuration for the Application Server.
@@ -38,6 +38,6 @@ var DefaultApplicationServerConfig = applicationserver.Config{
 		Timeout:   5 * time.Second,
 		QueueSize: 16,
 		Workers:   16,
-		Downlinks: webui.APIConfig{Enabled: true, BaseURL: shared.DefaultPublicURL + "/api/v3"},
+		Downlinks: web.DownlinksConfig{PublicAddress: shared.DefaultPublicURL + "/api/v3"},
 	},
 }